# backend/docs/swagger.yaml
basePath: /
definitions:
...
  api_rest.createFolderRequest:
    properties:
      name:
        type: string
      parentId:
        type: string
    required:
    - name
    type: object
...
  api_rest.moveFolderRequest:
    properties:
      targetParentId:
        type: string
    type: object
...
  api_rest.renameFolderRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://example.com/support
  description: REST, GraphQL, and WebSocket endpoints for the file service.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BalkanID File Service API
  version: "1.0"
paths:
...
  /api/v1/folders:
    get:
      consumes:
      - application/json
      description: List user's folders with optional parent filtering and trash support
      parameters:
      - description: Parent folder ID
        in: query
        name: parentId
        type: string
      - description: List deleted folders
        in: query
        name: deleted
        type: boolean
      - description: Limit results (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Folders list response
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List folders
      tags:
      - folders
    post:
      consumes:
      - application/json
      description: Create a new folder under optional parent
      parameters:
      - description: Folder creation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api_rest.createFolderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created folder response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create folder
      tags:
      - folders
  /api/v1/folders/{id}:
    delete:
      description: Permanently delete a folder and all its contents (files and subfolders)
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete confirmation
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete folder
      tags:
      - folders
    patch:
      consumes:
      - application/json
      description: Update the name of a folder
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      - description: Rename request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api_rest.renameFolderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rename confirmation
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Rename folder
      tags:
      - folders
  /api/v1/folders/{id}/ancestors:
    get:
      description: Retrieve the path of ancestor folders from root to the specified
        folder
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Array of ancestor folders
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get folder ancestors
      tags:
      - folders
  /api/v1/folders/{id}/contents:
    get:
      consumes:
      - application/json
      description: Get all files and subfolders within a folder
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Folder contents
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List folder contents
      tags:
      - folders
  /api/v1/folders/{id}/download:
    get:
      description: Download a folder and its contents as a ZIP archive
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/zip
      responses:
        "200":
          description: ZIP archive
          schema:
            type: file
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download folder as archive
      tags:
      - folders
  /api/v1/folders/{id}/files:
    get:
      description: Retrieve all files contained within a specific folder
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20)'
        in: query
        name: limit
        type: integer
      - description: Sort field (name, created_at, updated_at)
        in: query
        name: sort
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Array of files
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List files in folder
      tags:
      - folders
  /api/v1/folders/{id}/move:
    patch:
      consumes:
      - application/json
      description: Move a folder to a different parent folder
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      - description: Move request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api_rest.moveFolderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated folder information
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Move folder
      tags:
      - folders
  /api/v1/folders/{id}/tree:
    get:
      description: Retrieve the hierarchical tree structure of a folder and its subfolders
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Maximum depth to traverse (default: unlimited)'
        in: query
        name: depth
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Folder tree structure
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get folder tree structure
      tags:
      - folders
...
securityDefinitions:
  BearerAuth:
    description: 'Provide a Bearer token: "Bearer {access_token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"